# Generated Optimizations for KEVM