--- a/.gitignore
+++ b/.gitignore
@@ -10,3 +10,4 @@
 /tests/specs/**/*.prove.out
 /optimizer/*.k
 /optimizer/*.json
+/tests/specs/evm-optimizations-spec.md