--- a/optimizer/optimize.sh
+++ b/optimizer/optimize.sh
@@ -132,32 +132,36 @@ cp optimizer/optimizations.md optimizations.md
 git add optimizations.md
 git commit -m 'optimizations: clear optimizations' || true
 
-# score: (number with speedup >7%) - (number with slowdown >7%)
-doOptimization "PUSH(N)"  0 10 40                                                       #  73    65
-doOptimization "DUP(N)"   0 10 40 '#stackNeeded(DUP(N)) <=Int #sizeWordStack(WS)'       #  14    16
-doOptimization "SWAP(N)"  1 10 40 '#stackNeeded(SWAP(N)) <=Int #sizeWordStack(W0 : WS)' #  14    16
-doOptimization MSTORE     2 14 40                                                       # -18    11
-doOptimization ADD        2 10 40                                                       #  8     5
-doOptimization POP        1 10 40                                                       #  5     5
-doOptimization MSTORE8    2 14 40                                                       #  0     5
-doOptimization LT         2 10 40                                                       #  7     3
-doOptimization MOD        2 10 40                                                       #  0     3
-doOptimization AND        2 10 40                                                       #  0     2
-doOptimization EQ         2 10 40                                                       # -1     2
-doOptimization MUL        2 10 40                                                       # -1     1
-doOptimization NOT        1 10 40                                                       # -4     1
-doOptimization XOR        2 10 40                                                       # -9     0
-doOptimization SIGNEXTEND 2 10 40                                                       #  0     0
-doOptimization SGT        2 10 40                                                       #  0     0
-doOptimization GT         2 10 40                                                       #  0     0
-doOptimization SUB        2 10 40                                                       #  0     0
-doOptimization DIV        2 10 40                                                       #  0     0
-doOptimization ADDMOD     3 10 40                                                       # -1     0
-doOptimization SLT        2 10 40                                                       #  0    -1
-doOptimization EVMOR      2 10 40                                                       #  0    -1
-doOptimization ISZERO     1 10 40                                                       # -3    -3
-doOptimization MULMOD     3 10 40                                                       #  1    -4
-doOptimization MLOAD      1 14 40                                                       # -9    -4
+# score: (number with speedup >7%) - (number with slowdown >7%), for both (exec, spec)
+
+# were consistently positive
+doOptimization "PUSH(N)"  0 10 40                                                       #  73    65    (53  , 9 )
+doOptimization "DUP(N)"   0 10 40 '#stackNeeded(DUP(N)) <=Int #sizeWordStack(WS)'       #  14    16    (0   , 8 )
+doOptimization "SWAP(N)"  1 10 40 '#stackNeeded(SWAP(N)) <=Int #sizeWordStack(W0 : WS)' #  14    16    (13  , 2 )
+doOptimization ADD        2 10 40                                                       #  8     5     (39  , 9 )
+doOptimization SUB        2 10 40                                                       #  0     0     (18  , 4 )
+doOptimization AND        2 10 40                                                       #  0     2     (25  , 4 )
+
+# not sure it's worth it to include these ones
+# doOptimization MSTORE     2 14 40                                                       # -18    11    (-23 , -2)
+# doOptimization POP        1 10 40                                                       #  5     5     (-9  , 2 )
+# doOptimization MSTORE8    2 14 40                                                       #  0     5     (-15 , -9)
+# doOptimization LT         2 10 40                                                       #  7     3     (1   , 0 )
+# doOptimization MOD        2 10 40                                                       #  0     3     (-7  , 0 )
+# doOptimization EQ         2 10 40                                                       # -1     2     (-3  , 8 )
+# doOptimization MUL        2 10 40                                                       # -1     1     (-16 , -2)
+# doOptimization NOT        1 10 40                                                       # -4     1     (-10 , -8)
+# doOptimization XOR        2 10 40                                                       # -9     0     (-6  , 0 )
+# doOptimization SIGNEXTEND 2 10 40                                                       #  0     0     (9   , 0 )
+# doOptimization SGT        2 10 40                                                       #  0     0     (-22 , 0 )
+# doOptimization GT         2 10 40                                                       #  0     0     (12  , 2 )
+# doOptimization DIV        2 10 40                                                       #  0     0     (4   , 7 )
+# doOptimization ADDMOD     3 10 40                                                       # -1     0     (1   , 0 )
+# doOptimization SLT        2 10 40                                                       #  0    -1     (0   , 0 )
+# doOptimization EVMOR      2 10 40                                                       #  0    -1     (4   , 0 )
+# doOptimization ISZERO     1 10 40                                                       # -3    -3     (9   , 1 )
+# doOptimization MULMOD     3 10 40                                                       #  1    -4     (-4  , 0 )
+# doOptimization MLOAD      1 14 40                                                       # -9    -4     (14  , 6 )
 
 # spec requires multiple branches
 # doOptimization EXP 2 10